feat(core): tiered doc generation + graph metrics in SQL for large repos
Split two oversized hotspots into subpackages (every module <400 lines) and
add two scaling features for large codebases:

- Tiered doc generation: GenerationConfig.tier1_top_n caps full-LLM file
  pages; the long tail renders from a deterministic template + embedding
  (no LLM call). Default None keeps every selected page on the LLM path,
  so existing behaviour is unchanged unless the knob is set.
- Graph metrics in SQL: new graph_metrics table (alembic 0021) materialized
  after build, plus GraphBuilder.load_metrics_from_sql / release_graph so
  metric reads can be served from SQL without recomputing NetworkX
  centrality. Gated by OrchestratorMode.sql_backed_metrics (on in fast
  mode); standard generation keeps the in-memory graph for traversal.

CLI: surface fast mode in `repowise init` without a new top-level menu
option — a contextual offer when a large repo is detected, plus run-mode
and tier-cap knobs in advanced mode, and a fast-mode upgrade hint in the
next-steps panel.

Tests: tiering partition + no-LLM path, SQL metric correctness vs NetworkX,
CLI offer helpers, and Postgres integration tests for embed_batch and
graph_metrics (testcontainers, auto-skip without Docker).
