+++ packages/cli/src/repowise/cli/commands/init_cmd.py
@@ -1065,6 +1065,7 @@ def init_command(
         build_contextual_next_steps,
         format_elapsed,
         interactive_advanced_config,
+        interactive_fast_mode_offer,
         interactive_mode_select,
         interactive_provider_select,
         load_dotenv,
@@ -1073,6 +1074,7 @@ def init_command(
         print_phase_header,
         print_scan_summary,
         quick_repo_scan,
+        should_offer_fast_mode,
     )
 
     start = time.monotonic()
@@ -1148,6 +1150,10 @@ def init_command(
     # ---- Interactive mode (TTY, no explicit flags) ----
     is_interactive = sys.stdin.isatty() and provider_name is None and not index_only
 
+    # Tiered doc generation cap (set in advanced mode); None = every selected
+    # file page is a full-LLM tier-1 page (unchanged behaviour).
+    tier1_top_n: int | None = None
+
     # Pre-scan for interactive mode — fast stats to inform choices
     scan_info = None
     if is_interactive:
@@ -1159,9 +1165,18 @@ def init_command(
 
         if mode == "index_only":
             index_only = True
+            # On a large repo, an index-only run is exactly the case where the
+            # fast tier (essential git, no blame/co-change) pays off — offer it,
+            # defaulting to yes since docs are already opted out.
+            if (
+                run_mode != "fast"
+                and should_offer_fast_mode(scan_info)
+                and interactive_fast_mode_offer(console, scan_info, default_fast=True)
+            ):
+                run_mode = "fast"
         elif mode == "advanced":
             provider_name, model = interactive_provider_select(console, model, repo_path=repo_path)
-            adv = interactive_advanced_config(console, scan=scan_info)
+            adv = interactive_advanced_config(console, scan=scan_info, allow_fast=True)
             commit_limit = adv["commit_limit"]
             follow_renames = adv["follow_renames"]
             skip_tests = adv["skip_tests"]
@@ -1172,8 +1187,21 @@ def init_command(
             test_run = adv["test_run"]
             embedder_name = adv.get("embedder") or embedder_name
             include_submodules = adv.get("include_submodules", include_submodules)
+            run_mode = adv.get("run_mode", run_mode)
+            tier1_top_n = adv.get("tier1_top_n")
+            if run_mode == "fast":
+                index_only = True
         else:
             provider_name, model = interactive_provider_select(console, model, repo_path=repo_path)
+            # Full mode picked, but on a large repo offer the quick path too.
+            # Default no here — the user explicitly asked for docs.
+            if (
+                run_mode != "fast"
+                and should_offer_fast_mode(scan_info)
+                and interactive_fast_mode_offer(console, scan_info, default_fast=False)
+            ):
+                run_mode = "fast"
+                index_only = True
 
     editor_options = resolve_editor_setup_options(
         console,
@@ -1407,6 +1435,7 @@ def init_command(
             language=language,
             reasoning=resolved_reasoning,
             enable_onboarding=onboarding,
+            tier1_top_n=tier1_top_n,
         )
         if sys.stdin.isatty() and coverage_pct is None and not yes:
             options = compute_coverage_options(
@@ -1816,6 +1845,7 @@ def init_command(
 
         next_steps = build_contextual_next_steps(
             index_only=True,
+            fast_mode=(run_mode == "fast"),
             dead_unreachable=_dc_unreachable,
             dead_unused=_dc_unused,
             hotspot_count=_hotspot_count_final,
