+++ packages/core/src/repowise/core/generation/models.py
@@ -133,6 +133,15 @@ class GenerationConfig:
     jobs_dir: str = ".repowise/jobs"
     large_file_source_pct: float = 0.4  # use structural summary when source tokens > budget * this
     language: str = "en"
+    # ---- Tiered doc generation (large-repo scale) ---------------------
+    # Caps the number of file pages that receive full LLM generation.
+    # The top ``tier1_top_n`` selected file pages by PageRank are
+    # generated by the LLM (tier-1, the existing path); the remaining
+    # selected file pages are rendered from a deterministic Jinja
+    # template and embedded for search — with no LLM call (tier-2).
+    # ``None`` (default) preserves the prior behaviour exactly: every
+    # selected file page is a full tier-1 LLM page.
+    tier1_top_n: int | None = None
 
     def __post_init__(self) -> None:
         if self.embed_concurrency is None:
@@ -186,7 +195,7 @@ class GeneratedPage:
     confidence: float = 1.0
     freshness_status: str = "fresh"  # FreshnessStatus literal
     metadata: dict[str, object] = field(default_factory=dict)
-    # 1–3 sentence purpose blurb extracted from the rendered content. Used by
+    # 1-3 sentence purpose blurb extracted from the rendered content. Used by
     # MCP get_context as the default narrative payload (content is gated behind
     # include=["full_doc"]).
     summary: str = ""
