+++ packages/core/src/repowise/core/persistence/__init__.py
@@ -18,12 +18,11 @@ from repowise.core.providers.embedding.base import Embedder, MockEmbedder
 
 from .crud import (
     batch_upsert_graph_edges,
+    batch_upsert_graph_metrics,
     batch_upsert_graph_nodes,
     batch_upsert_symbols,
     bulk_upsert_decisions,
     bulk_upsert_external_systems,
-    link_graph_nodes_to_external_systems,
-    list_external_systems,
     count_chat_messages,
     create_chat_message,
     create_conversation,
@@ -43,6 +42,7 @@ from .crud import (
     get_git_metadata,
     get_git_metadata_bulk,
     get_graph_edges_for_node,
+    get_graph_metrics,
     get_graph_node,
     get_graph_nodes_by_ids,
     get_node_degree_counts,
@@ -53,9 +53,11 @@ from .crud import (
     get_stale_decisions,
     get_stale_pages,
     get_top_entry_points,
+    link_graph_nodes_to_external_systems,
     list_chat_messages,
     list_conversations,
     list_decisions,
+    list_external_systems,
     list_page_ids,
     list_pages,
     load_prior_pages,
@@ -99,6 +101,7 @@ from .models import (
     GenerationJob,
     GitMetadata,
     GraphEdge,
+    GraphMetric,
     GraphNode,
     Page,
     PageVersion,
@@ -132,6 +135,7 @@ __all__ = [
     "GenerationJob",
     "GitMetadata",
     "GraphEdge",
+    "GraphMetric",
     "GraphNode",
     # vector store
     "InMemoryVectorStore",
@@ -148,25 +152,15 @@ __all__ = [
     "async_sessionmaker",
     # crud
     "batch_upsert_graph_edges",
+    "batch_upsert_graph_metrics",
     "batch_upsert_graph_nodes",
     "batch_upsert_symbols",
     # decision crud
     "bulk_upsert_decisions",
     # external systems crud (C4 L1)
     "bulk_upsert_external_systems",
-    "link_graph_nodes_to_external_systems",
-    "list_external_systems",
     # chat crud
     "count_chat_messages",
-    # graph read-side queries
-    "get_all_file_metrics",
-    "get_community_members",
-    "get_cross_community_edges",
-    "get_graph_edges_for_node",
-    "get_graph_node",
-    "get_graph_nodes_by_ids",
-    "get_node_degree_counts",
-    "get_top_entry_points",
     "create_chat_message",
     "create_conversation",
     "create_engine",
@@ -174,10 +168,14 @@ __all__ = [
     "delete_conversation",
     "delete_decision",
     "delete_repository",
+    # graph read-side queries
+    "get_all_file_metrics",
     # git metadata crud
     "get_all_git_metadata",
+    "get_community_members",
     "get_configured_db_url",
     "get_conversation",
+    "get_cross_community_edges",
     "get_db_url",
     # dead code crud
     "get_dead_code_findings",
@@ -187,6 +185,11 @@ __all__ = [
     "get_generation_job",
     "get_git_metadata",
     "get_git_metadata_bulk",
+    "get_graph_edges_for_node",
+    "get_graph_metrics",
+    "get_graph_node",
+    "get_graph_nodes_by_ids",
+    "get_node_degree_counts",
     "get_page",
     "get_page_versions",
     "get_repo_db_path",
@@ -195,10 +198,13 @@ __all__ = [
     "get_session",
     "get_stale_decisions",
     "get_stale_pages",
+    "get_top_entry_points",
     "init_db",
+    "link_graph_nodes_to_external_systems",
     "list_chat_messages",
     "list_conversations",
     "list_decisions",
+    "list_external_systems",
     "list_page_ids",
     "list_pages",
     "load_prior_pages",
